Complete and use new NetCoreShim project
