--- source/Calamari.AzureWebApp.DotNetCoreShim/Calamari.AzureWebApp.DotNetCoreShim.csproj
@@ -1,19 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>net462</TargetFramework>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="CommandLineParser" Version="2.9.1" />
-      <PackageReference Include="Microsoft.Web.Deployment" Version="4.0.5" />
-      <PackageReference Include="Serilog" Version="4.0.2" />
-      <PackageReference Include="Serilog.Sinks.Console" Version="6.0.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Calamari.Common\Calamari.Common.csproj" />
-    </ItemGroup>
-
-</Project>
